requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deep-translator==1.11.4
distro==1.9.0
Django==5.0.7
frozenlist==1.4.1
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==1.0.5
httpx==0.27.0
hyperframe==5.2.0
idna==2.10
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-core==0.2.25
langchain-groq==0.1.8
langchain-text-splitters==0.2.2
langsmith==0.1.94
mangum==0.17.0
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
tenacity==8.5.0
textblob==0.18.0.post0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
yarl==1.9.4
